Refactor evictor/mod.rs to evictor.r
